removed old align pipeline
