+++ nail/src/cli.rs
@@ -17,8 +17,6 @@ pub struct CommonArgs {
 pub enum SubCommands {
     #[command(about = "Run the entire nail pipeline: prep, seed, & align")]
     Search(SearchArgs),
-    #[command(about = "Run the entire nail pipeline: prep, seed, & align")]
-    Search2(SearchArgs),
     #[command(about = "Prepare a query (MSA) file and target (fasta) file for the seed step")]
     Prep(PrepArgs),
     #[command(about = "Use MMseqs2 to create a set of alignment seeds for the align step")]
